pyproject.toml
[tool.poetry]
name = "finalproject-gafarov-m25-555"
version = "0.1.0"
description = "Валютный кошелек для торговли валютами"
authors = ["Гафаров <gafarov@example.com>"]
readme = "README.md"
packages = [{include = "valutatrade_hub"}]

[tool.poetry.dependencies]
python = "^3.9"
prettytable = "^3.10.0"
requests = "^2.31.0"  # Для работы с API
python-dotenv = "^1.0.0"  # Для загрузки переменных окружения

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[tool.poetry.scripts]
project = "valutatrade_hub.cli.interface:main"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
  "E",
  "W",
  "F",
  "I",
  "B",
]

[tool.ruff.format]
quote-style = "double"

[tool.valutatrade]
data_dir = "data"
rates_ttl_seconds = 300  # 5 минут
base_currency = "USD"
log_format = "json"
log_file = "logs/actions.log"

[tool.valutatrade.parser]
exchange_rate_api_key = "${EXCHANGERATE_API_KEY}"
request_timeout = 10
fiat_currencies = ["EUR", "GBP", "RUB", "JPY", "CNY"]
crypto_currencies = ["BTC", "ETH", "SOL", "BNB", "ADA"]